Top5Reach update
The computation of the pss’s with the most reach (Top5Reach) now doesn’t take into account the column pss on the opinion’s table. Instead it searches for all the pss’s on the existing models and the associated pairs source/account and calculates the average reach of the opinions that correspond to this pair. Also, the computation related to the contruction of the top 5 pss chart in the index page is now made only if the request is different of the previous requests (for that a new column – request VARCHAR(200) -  was added to the table reach). This table is cleared every time oversight runs.